backbone: refactor CoupledL2 top-level framework
TL2TLCoupledL2 and TL2CHICoupledL2 used to flatten all the logics in
each top-level module, which makes two versions of CoupledL2 repeat lots
of the logics. In fact, these two versions share most of the logics in
the top level except for outer interface (one for TileLink and one for
CHI).

This commit extracts the common parts between TileLink and CHI and
modularizes into CoupledL2, so that TL2TLCoupledL2 and TL2CHICoupledL2
only need to implement the differences with each other.
